Enkelt körs som vanligt från enkelt.py men skripter kör nu ett annat skript (run_enkelt.py) uti sig själv, detta gör för att b.la. kunna byta ut True -> Sant och False -> Falskt
